Do the check in calculate-versions instead of in publish-packages
